move `type_check` out of `transform` and into the `nll` module
