+++ src/librustc_mir/borrow_check/nll/subtype_constraint_generation.rs
@@ -14,11 +14,11 @@ use rustc::infer::region_constraints::RegionConstraintData;
 use rustc::infer::region_constraints::{Verify, VerifyBound};
 use rustc::ty;
 use syntax::codemap::Span;
-use transform::type_check::Locations;
-use transform::type_check::MirTypeckRegionConstraints;
-use transform::type_check::OutlivesSet;
 
 use super::region_infer::{TypeTest, RegionInferenceContext, RegionTest};
+use super::type_check::Locations;
+use super::type_check::MirTypeckRegionConstraints;
+use super::type_check::OutlivesSet;
 
 /// When the MIR type-checker executes, it validates all the types in
 /// the MIR, and in the process generates a set of constraints that
